drbdmanageNG Controller & Security model initialization
The security model's core classes are ObjectProtection
and AccessContext.

The ObjectProtection class is used to check access permission
to objects. The AccessContext class represents a subject that
operates upon such protected objects.

The security model comprises
- Identities (basically, user ids)
- Roles (a user's role)
- Access control lists
- Security types & domains
- Privilege sets

ObjectProtection instances manage
- an object's creator Identity
- an object's owner Role
- the object's Access control list
- the object's Security type

AccessControl instances manage
- a subject's Identity
- a subject's Role
- a subject's Security domain
- a subject's Privilege sets

An AccessContext instance can be created by logging in with valid
credentials (not implemented yet), or by impersonation of the
specified Identity, Role and Type performed by a
subject operating with all privileges.

An AccessContext instance can be cloned, e.g. to obtain multiple
AccessContext instances for multiple threads, where it is required
that each thread can change its AccessContext independently, e.g.
by enabling or disabling privileges.

An ObjectProtection instance can be created from an AccessContext.
Upon creation, the ObjectProtection instance's creator Identity,
owner Role and Security type are initialized from the AccessContext
instance's subject Identity, subject Role and subject Security domain,
respectively.

An Identity identifies the subject that is performing an action.
Upon creation of a new protected object, the Identity becomes
the creator of that object.

A Role identifies the subject's role under which an action is
performed. The Role typically determines the level of access
that a subject has to protected objects, by means of
Access control list checks. Initial privileges are typically
determined by the Role that has been logged in, but may be
changed afterwards.
Upon creation of a new protected object, the Role becomes the
owner of that object.

Access control lists implement discretionary access control (DAC).
Access control lists determine the level of access that a role
has to a protected object. An Access control list can be changed
by the owner of the object, or by a role that has been granted
CONTROL access to the object by an access control entry in the
Access control list, or by a subject that operates with
the PRIV_OBJ_OWNER privilege.

Security types & domains implement mandatory access control (MAC).
Security types specify the type of protected objects and the domain
of a subject. Access is allowed if there is a rule that explicitly
allows the level of access by the subject's domain to the
protected object's type, or if the subject is operating with the
PRIV_MAC_OVRD privilege.
Upon creation of a new protected object, the security type of that
object will typically be the same as the security domain of the
subject that created the object.
The security type of objects and the security domain of subjects
can not normally be changed arbitrarily, except by a subject
operating with all privileges.

Privilege sets manage the privileges that are granted to a
subject's current AccessContext.
There are two privilege sets, the effective privilege set and the
limit privilege set.
The effective privilege set determines the privileges that apply
to actions performed by the subject. The limit privilege set
determines the upper bound of effective privileges that can be
enabled in the effective privilege set.
Upon creation of an AccessContext instance, the limit privilege
set is determined by the Role that the subject has been
logged in to, or, if the AccessContext is created by impersonation
by a subject operating with all privileges, it can be
set arbitrarily during impersonation.
The effective privilege set is initially empty. Privileges
contained in the limit privilege set can be enabled in the
effective privilege set. Any enabled privilege can be disabled.
Privileges in the limit privilege set can disabled, which will
effectively also disable the privilege in the effective privilege
set. The limit privilege set can only be reduced, but
can never be extended.

The system's AccessContext instance, which is required for
initializing objects, identities, roles, types, domains, etc.,
during system start and while loading objects from persistent
storage, is created during startup by the Initializer class,
and is then used to create the drbdmanageNG instance.