Split debug command superclasses for future Controller/Satellite debug generalization
The Controller and Satellite modules share many of their features such as
network communications and service threads, while other features are unique
to each module. Therefore, some debug commands may have a common
implementation for the Controller and the Satellite, while others will need to
be specifically designed for debugging either the Controller or the Satellite.

Changes are:
- CommonDebugControl is the debug interface for accessing functionality
  that the Controller and the Satellite share
- Controller.DebugControl is the debug interface for
  controller specific functionality; extends CommonDebugControl
- Satellite.DebugControl is the debug interface for
  satellite specific functionality; extends CommonDebugControl
- CommonDebugCmd is the interface for all debug commands
- BaseDebugCmd is an abstract class that handles access to a debug
  command's alias names, parameter descriptions, etc., as well as
  generic debug functionality; implements CommonDebugCmd
- ControllerDebugCmd is the interface for controller specific
  debug commands
- SatelliteDebugCmd is the interface for satellite specific
  debug commands
- BaseControllerDebugCmd is an abstract class that handles
  initialization of controller specific debug commands
- BaseSatelliteDebugCmd is an abstract class that handles
  initialization of satellite specific debug commands

Controller specific debug commands are supposed to 
extend BaseControllerDebugCmd
(and thereby implement ControllerDebugCmd)

Satellite specific debug commands are supposed to
extend BaseSatelliteDebugCmd
(and thereby implement SatelliteDebugCmd)

Generic debug commands that are shared by the Controller
and the Satellite are supposed to extend BaseDebugCmd
only (and thereby implement CommonDebugCmd).