Changed queryAccess() -> getEntry(), reimplemented queryAccess()
The queryAccess() method can now be used to query the level of access
granted to a protected object to a security context. The getEntry() method
can be used to query the level of access granted to a protected object
by a specific security component, without taking the current security
level or any enabled privileges into account.

In other words, e.g. applied to an access control list,
queryAccess() will determine the currently effective level of
permissible access, while getEntry() will determine the level of access
that is granted by an entry in the access control list.