Reworked usage of transMgr in the DbDrivers
transMgr is no longer null-checked when passed to DbDrivers. Satellites do not use transMgr,
so no NullPointerException possible. Controller use them and will cause a
NullPointerException on purpose, as otherwise the data would have gotten updaten in RAM but
not in the DB (thus, ImplementationError), which is caused by simply forgetting to
call .setConnection(transMgr) before the change.

Additional work:
* Simplified the .getInstance() methods of the data classes
* Db-load methods now only log warnings when the requested entry is missing if the
  logWarnIfNotExists boolean is set (this prevents false-positive warning messages when
  a .getInstance() initially asks the empty db about the existence, and afterwards
  creates the object successfully)
* SatelliteDbDriver now loads objects from their parents. (e.g a queried resource is loaded
  from its parent node where it had to be placed already on creation) This is needed to
  ensure consistent references between the data objects