Changed access required to create or load an ObjectProtection instance to CHANGE.
There are basically two cases for the creation of ObjectProtection instances:
- Loading an existing one from the database
- Creating a new instance for a newly created object
  (e.g., an ObjectProtection instance for a newly created NodeData object)

The first case normally occurs only during startup or configuration reload
and is always executed in the system context (and there may be some
check in place to ensure that only a privileged context loads instances
from the database).

The second case occurs when any access context creates the ObjectProtection
for some new object, and in this case, CHANGE access is considered
sufficient for object creation. While the creator of an object normally has
CONTROL access, mandatory access control rules may
limit access for the creator to CHANGE.