Controller debug-logs now unknown apis from sat
If a client sends an unknown API-call the response is unchanged.

As the satellite does not have to know about the controller as a node,
it cannot distinguish between controller and client. That means, if the
controller sends the satellite an unknown API-call, the satellite will
answer him as it would be a client.

The controller can distinguish between client and satellite, so it will
prevent an endless "unknown-api-call" ping pong by just logging with
debug level and not answering the satellite.