TcpPortNumber: Set minimum port number to 1
Port number 0 is normally only valid when opening a client socket, where
it is used to specify dynamic allocation of any free port number.
If this is ever required in Linstor, dynamic allocation can be indicated
by either some additional piece of information or by setting the
reference to an otherwise specified TcpPortNumber to null.