Resource: Made the Node and ResourceDefinition reference updatable
On the satellite, when a Node object or ResourceDefinition object is
updated by replacing the instance, all associated Resource objects
must be updated to reference the new Node or ResourceDefinition object.
This is still missing in the code and was impossible before, because
the references in the Resource object were declared final.

Added two methods:
setAssignedNode() - for updating the Node reference
setDefinition() - for updating the ResourceDefinition reference

Both methods check the access context for SYSTEM identity to prevent
inadvertent use by APIs or debug commands using a non-SYSTEM
access context.