Node, Resource, ResourceDefinition: Added runtime instance UUIDs for debugging
The idea is that every instance created at runtime is assigned a
UUID, so that if different instances of an object that should be
unique are present, this can be detected by debug code.
E.g., if a Node object with a NodeName of 'AlphaCentauri' exists in the
controller's or satellite's nodes map, and a Resource object references
a Node with the NodeName of 'AlphaCentauri', then if those Node objects
have different runtime instance UUIDs, either the nodes map or the
Resource object has an incorrect reference to an old version of the
Node object.