Moved Volume.Key -> VolumeDefinition.Key
The static nested class Key fits better into the VolumeDefinition
class, because it only contains enough information to
uniquely identify a VolumeDefinition of a ResourceDefinition,
but not a Volume of a Resource.

The satellite uses the key for unique identification of
Volumes of multiple Resources in one set, because it
only deals with a single Node object, which represents
the satellite's local node.
Therefore, no collisions between the volume keys of
multiple resources on multiple nodes are possible and
the node name can remain absent from the key.