build: generate release tarballs
This allows us to generated release tarballs as usual (make
prepare_release/debrelease). It includes some checks to ensure the tree
is in a sane state.

What you do before a release:
- set "version" in build.gradle (e.g., 0.1)
- commit everything
- tag HEAD as v$version (e.g., v0.1)
- make prepare_release

make debrelease will be the same, but these tarballs include the
"debian" directory. That dir will exist eventually.