Removed Transaction manager from respondeResource
The method aquires only read-lock on the corresponding maps, yet
got the node instance with .getInstance, which in return tried to set
the given transMgr which caused an attempt of overriding a possible
existing transMgr.

This way the method does not "activate" the node (thus accesses the node
in a read-only way), which prevents the problem.