Drbdadm primary will now be called after creating a resource
The handshake works as follows:
After creating the resource on the satellite a request primary message
will be send to the controller if the primary property is not set
on the resource definition.
The controller will check if the resource definition still doesn't have
the property and choose one of it's resources to be primary and
send a set primary message to this satellite.
The satellite will then fetch the resource and send a volatile flag
to mark it primary and the devmgr will call drbadm primary with the resource.