Satellite applies updates now ordered
In detail this works as follows:
If the authentication message of the controller is successful, the
response now also contains an expected fullSyncId. Here the satellite
drops all references to its local nodes, rscDfns and storPoolDfns, in
other words: to every local LinStor-Object. The localNode is recreated.

The controller next sends the fullSync with the satellite's fullSyncId.
Doing so, the controller also stores the fullSyncId at the peer instance,
which additionally keeps track of the updateId (starting by 0 with each
peer).
The satellite only applies a fullSync if it matches the expected
fullSyncId.

When any change happens, the controller notifies the satellite. This
works as before, no additional ids or synchronization.
Also the requests from the satellite did not change.
However, serializing data for node, resource or storage pools on the
controller increases the peer's updateId (AtomicLong). The serialized
data objects are extended with the current fullSyncId and their unique
(within the current fullSyncId) updateId.
If the satellite receives an update not matching the "next" updateId,
the update is stored. If an update with matching updateId arrives, that
update is applied and the stored updates are checked if the next higher
updateId is already stored or not.