Added configuration files for automated checks of code compliance
- LINSTOR_Coding_Standard_checkstyle.xml
  configures the checkstyle utility for automated checks of the
  source code for compliance with the Coding Standard used in
  the LINSTOR project

- LINSTOR_Code_Optimization_checkstyle.xml
  configures the checkstyle utility to assist with finding code
  sections that may be too complex, or that contain too many
  duplicates of string literals, etc.

Note that the checks are not perfect, so they will not detect
every piece of code that is incorrect, implemented in a bad
way or formatted incorrectly, and may also be false
positives sometimes.
The automated checks are not supposed to be a
replacement for a comprehensive review and
quality assurance process.