Introduced AccessToDeletedObjectException
Also replaced all corresponding ImplementationErrors with this new
exception.

Additionally fixed a bug with TransactionSimpleObjects. The TSO
did not forward the TransactionManager to its managed object.
Simply forwarding the object caused an infinite recursion, which
is prevented by asking the TransactionManager if the Transaction-
Object we currently want to register is already registered.