Introduced a separate thread pool for satellite services
The DeviceManager service schedules device handlers while holding
the reconfigurationLock as a Reader and waits for the device handlers
to finish.

If the NetCom services schedule API execution on the same thread pool
as the one used by the DeviceManager service, then APIs that
lock the reconfigurationLock as a Writer, such as e.g. the
Controller<->Satellite FullSync API, will block, and if all of the
thread pool's threads are in use for such APIs, then the DeviceManager
will either block on the WorkQueues Semaphore lock because the
WorkQueue is full, or it will wait forever for the tasks that it placed
in the WorkQueue to finish, because no threads are available to execute
those tasks. In both cases, both the DeviceManager and the NetCom
services will deadlock.

Using a separate work queue for the device manager avoids this problem
by ensuring that the device manager will finish, and as a result,
will release the reconfigurationLock, which lets the blocked APIs
finish as well.