Fixing "nested transaction seem not to work with in memory dbs"
This reverts commit 06f1f23afa0521fa4582eef177e0d6fd3cd29773.

The following happened:
Not InMemory database connections had their username set to
linstor, which is also the SCHEMA for the linstor database.
If a connection does not specify a SCHEMA, the username is taken
instead.
This is why the non-in-memory database worked properly.

By using an in-memory database, until now we did not specify
neither a user nor a SCHEMA.
However, the first connection created the whole database, and
one of its first statements was the SET SCHEMA 'LINSTOR';
If the connection gets closed, it can be reused for the next
statement, but it remembered its SCHEMA.
The first "nested" connection, however, had no SCHEMA defined,
which defaulted in some db-specific default SCHEMA. Obviously
this default SCHEMA did not contain any of our tables or view
and resulted in an SQLException.