Large rework of TransactionMgr
The transaction manager does not need to be added manually anymore.
The CommonMessageProcessor seeds the current api-scope with the
"current" TransactionMgr instance, and the database-drivers fetch
that instance via Provider<TransactionMgr> getters.

For the in-memory part, every TransactionObject (NodeData,
TransactionSimpleObject, ... ) automatically call a setConnection(..)
with the "current" TransactionMgr instance when they are modified.

This basically means that the whole API layer does not have to care
about activating objects anymore. They only have to perform a commit
when they successfully finish their work. Otherwise the
AbsApiCallHandler will rollback the current transaction.

Operations outside of the AbsApiCallHandler or even outside of the
CommonMessageProcessor will have to enter seed the corresponding
LinstorScope with the current TransactionMgr and exit after the scope
after the operation. Otherwise the TransactionObjects and the database
drivers will throw NullPointerExceptions.