Added cryptsetup feature to storage driver
NOTE:
The crypt-password is currently stored as a property and will be
listed as an ordinary property. This might get blacklisted in future
but is currently left visible on purpose for easier debugging.

It is also possible to change that property, but this might
be a very bad idea as the LUKS container will not be recreated but
made unusable.

If the VlmDfnFlags.ENCRYPTED flag is set during creating a volume
definition, the controller will generate a random 20 character
long password and setup LUKS container using cryptsetup.

The drbd .res file will contain that crypt-device as blockdevice,
which means that drbd will replicate unencrypted data.

When using encrypted storage drivers the underlaying LVM cannot be
deleted while the LUKS container is still open. This might be
noteworthy when searching cleanup-bugs in automated tests.