Limit device manager worker thread count
Most of the external utilities do not scale well, to running a
greater number of them concurrently only increases memory
usage and congestion, but does not make the system faster.
For this reason, limit the maximum number of
device manager worker threads to much less than the
maximum number of Linstor main worker threads.