Fixed Base64 ArrayIndexOutOfBoundsException: Replaced by Java 8 class library implementation
The Base64 class encountered an ArrayIndexOutOfBoundsException trying to parse a
string that contained white space.
Since the project has been migrated to the Java 8 platform, rather than fixing the bug
in the custom implementation of the Base64 codec, I replaced the implementation
with the Base64 codec from the Java 8 class library.