Fixed deleting a node which has to cleanup first
Prior this fix the node got marked deleted, and also the corresponding
resources got marked deleted, but only the node-update was sent to the
satellites, not the changed resources.

Additionally, when the satellite finished deleting a resource, it
notifies the controller and sending the new storage pool free spaces.
However, that updates of the free spaces might try to access the just
deleted node, which is why this is aditionally checked in the
updateFreeSpace method.