Fix incorrect checking of zfs extend size [BREAKS old ZFS]
volblocksize is the correct extend size indicator.
This volblocksize can be different on any subvolume of a zfs pool
so there is no global setting like in LVM.
There we have to query existing subvols for their volblocksize,
new sub volumes are generated using the default 8K.

This breaks compatibility to already created zfs volumes,
as they where created with the wrong extend size.