move from /opt to /usr/share
This was initially triggered by packaging for the PPA. lintian throws
errors when you install files to /opt. And it complained that the jars
are not in /usr/share.

After discussing that with Phil and Robert, we decided to put the files
where they belong, which is for most of them (the jars) /usr/share.

This commit changes the default location, but needs some follow-up work.
The controller should accept some further arguments:
--db: path to the DB (e.g., /var/lib/linstor-server/)
--logs: path to the logs location (e.g., /var/log/linstor-server/)
--conf: path to the configuration (e.g., /etc/)
Maybe for --db and --logs the last component is optional, and
"linstor-server" is hard coded.

These should then be added to the service file.

Signed-off-by: Roland Kammerer <roland.kammerer@linbit.com>
Signed-off-by: Philipp Reisner <philipp.reisner@linbit.com>
Signed-off-by: Robert Altnoeder <robert.altnoeder@linbit.com>