Added exception handling comment, slightly optimized value parsing
Changed Short.valueOf() to Short.parseShort(). The valueOf() method
cashes some Short objects instead of creating new ones, however,
the object is discarded anyway in this statement when the value
is unboxed. Short.parseShort() returns the value as the basic
datatype short instead, which is also what valueOf() uses
internally.