build: add "release" target
"prepare_release" was how it is called in utils and the kernel part. All
other projects have a "release" target (DM, linstor-client, python-API).

"release" is the one that does not add the debian/ directory. This is
the one that should be used to publish "source" tarballs, and this
target is also used to import the resulting tarball in the PPA repos.
The PPA repos have their own debian/ directory per Debian release
branch.