Fix: LvmDriver
Commit `fa5d44e939839bd119273a817a51bb0cf2c066bc` introduced a bug where
deleteVolume looks for the volume which should be deleted. In case the
volume could not be created in the first place, that lookup was
configured to throw an exception because it didnt find the volume.