Made AutoPlace idempotent.
When a resource is already deployed X times and a new autoPlace is called
with placeCount:
* > X: usual case (X might be 0). Default success RC is returned
* = X: WARN_ALREADY_DEPLOYED RC is returned
* < X: FAIL_INVLD_PLACE_COUNT

Additionally, the user can first autoPlace a resource X times, and
afterwards call autoPlace for the same resource X+Y times for creating Y
new resources on a possibly different storage pool. However, be aware that
you always have to specify the target count of deployed resources, not the
"additional" count (that means X+Y instead of only Y).