Fixed hung resource creation
When the satellite creates a new resource, it is possible that this
resource already exists exactly as needed. In this case DRBD events2
stream will not contain a created resource (only an exists) event.

For fullSync, the satellite already handled this scenario. This commit
uses the same checking mechanism but for regular create resources
(after a completed fullSync).