Collect real free space information from LVM thin pools
The information is only updated when changes are made or the controller
reconnects to the satellite. However the free space in a thin pool
changes as it is used. Hence we do not show these values to the user.
They are however used for auto-placement because using out-of-date data
is better than always assuming that all thin pools have the same free
space.