Added shared StorPools
Shared storage pools are handled by the FreeSpaceManager. One
FreeSpaceManager can handle one or more storpools.

The user can specify a --shared-space name for a StorPool. All
StorPools with the same shared-space-name have the same
FreeSpaceManager instance.

When the Controller creates a Volume, that Volume is in an "adding"
state for the FreeSpaceManager. In this state, the FreeSpaceManager
still knows the last reported exact free space, but is also able to
calculate an estimation including the Volume(s) currently in the
"adding" state. These two values can differ as some StorageDrivers
(like LVM) have an extent size, which will round up the requested
size.

Once the Satellite finished deploying the Volume(s), the reported
success event to the Controller contains the new exact free space.
With this event the FreeSpaceManager removes the Volume from its
"adding" state.

Removing a Volume has the same procedure with a seperate
state-tracking.