Swordfish: driver can now continue with half-created volume
When a volume was create but not attached and the satellite stopped
working the satellite thought at the next start that the swordfish volume
was successfully created and run into an exception when trying to
determine the local nvme device path.

The "volumeExists" check was extended by also checking if the volume is
also attached, if not, the "createVolume" first checks if the volume was
already created (if not create it) and continues with attaching
afterwards.