StorageDrivers can now write into VolumeDefinition's Props
Some storage drivers generate data while creating the volume
(for example the swordfish driver). This data could be needed by other
volumes of the same volume definition.

Most methods of the StorageDriver interface are now getting the
(root-) PropsContainer of the corresponding VolumeDefinition.

In the internal "notifyResourceCreated" API message, those
VolumeDefinition properties are also included and stored in the
controller's database.

Be aware that multiple StorageDrivers concurrently generating data will
cause a race-condition when reporting that data back to the controller!