Fixed access to deleted volume
When the creation of a resource / volume failed and the controller never
received the NotifyResourceCreated event, the FreeSpaceMgr never deletes
its reference to the volume. When the volume is somehow deleted later, the
next access to "getFreeSpaceCurrentEstimation" will access that (deleted)
volume, resulting in a Error "Access to deleted volume".

This commit declares the volume as "created" when it actually gets removed
from the storage pool. This sounds contradictory, but can only do two
things:
Either the volume is still in the FreeSpaceMgr's list where it should be
removed (declaring the volume as "created" removes it from that list).
Or the volume was correctly created, meaning that at the time of deleting
the volume from the storPool it is no longer in the FreeSpaceMgr's list.
Removing the volume from a list it is not contained in results in a no-op.
Both cases are fine and the updated free space is correct and can safely
be applied.