Added swordfish_target node creation API
Creating a new node with type swordfish_target now spawns a new
satellite-process running on "localhost" and a automatically chosen port.
This process will be shutdown if the node is deleted or if the controller
itself shuts down.
If the chosen port is already in use, the creation is rollbacked and
retried with the next non-occupied (known by linstor) port.

If the controller starts up again it tries starting the known
swordfish_target nodes.
If the persisted port is already in use, a warning is logged reminding the
user that the port can be changed with modifyNetIf.

A swordfish_target must have exactly one network interface.

If the port if the network interface changes, the controller tries to stop
the process (if running) and spawns a new process with the new port (or
fails and aborts the modifyNetIf if the port is already in use)