Notify connection listeners after disconnecting a deleted node
This fixes a bug where a resource definition delete and node delete that
were both started while the node was offline led to the resource
definition not being deleted. The problem was that the responses
corresponding to the deletion of the resource for the resource
definition deletion were dropped because the node was disconnected as
part of the node deletion. With this fix, the resource definition
deletion is restarted after the resource has been deleted, so it
completes without needing to do any more satellite updates.