Only initiate an initial sync when the DRBD meta data is new
This prevents linstor from starting a sync on an existing resource. For
instance, in the case of migrations. Meta data is defined to be new if
we just created it or the current generation ID has a special value.
The DrbdPrimarySetOn property is still set in this situation to indicate
that the resource has at some point been synced. This prevents other
resources from starting a sync.