Fixed possible "Access to deleted <object>"
The problem was caused by the three {Node,Resource,Volume}Connection maps
as those had the actual Node, Resource or Volume as a key in their
TreeMap.
Whenever such a Node, Resource or Volume was deleted, the deletion first
deleted all of its *Connection entries. If for whatever reasons that
deletion had to be rolled back, the TreeMap had to be re-filled with the
actual *Connection entries. If this his re-insertations was made
before the deleted-flag was rolled back to false, the TreeMap compared
the re-instered data with existing entries, thus accessing the Node's,
Resource's or Volume's data (as implemented in their compareTo method).
This triggered the mentioned "Access to deleted *" exception.

This fix simple replaces the Node,Resource and Volume key-types for the
map with immutable NodeName, Resource.Key and Volume.Key types. Those can
be access even if the delete flag is (still) set to true.