Make new node connection a little less racy
This fixes an issue where the controller could fail to immediately send
authentication to a newly connected satellite. This then led to other
APIs sending messages to the satellite before the auth message. The
satellite responded to these messages with an error, as expected. The
issue was that the satellite could connect between the initial check for
whether it had connected and it being added to the reconnector list.

Also remove an unnecessary set of the node peer, which is performed by
the connector service.