Swordfish: fixed state-checking issues
One of them was that the initiator-driver assumed that exactly one
Endpoint exists. This restriction had to be loosen in order to check the
current attaching state.

This rework assumes that even with multiple Endpoints, all "Target"
ConnectedEntities have the same nqn (as they should point to the same
target), thus only the first Target's nqn is considered.