Swordfish_Target nodetype does no longer try to move linstor_common.conf
This behaviour resulted in annoying error reports but had no effect on the
functionality. Swordfish_Target should only run on the controller node. If
that node wants to perform DRBD-related actions, it has to have a "normal"
satellite running on the same machine. This "normal" satellite would
perform the generation and move of linstor_common.conf.
The controller AND Swordfish_Target types do not have to do this.