Swordfish: Exclude specially handled status code
In the case of DetatchResource, a 400 response code was first checked if
it contains an error message stating something like "the resource is not
attached". In this case, linstor continues without error.
In every other case of a 400 linstor throws the 400 as an exception.
However, the generated ErrorReport then contains something like
"Expected response codes: [200, 202, 400], Error response code: 400"
which is quite confusing.