Fix: skip storPool-size-check in case of migration
The fix simply skips the size-check if the volume-definition has set the
migration specific property. If the volume does not exist on the
satellite, it will be created if possible. If the volume does not exist
and the migration-property is set, the satellite might fail because of
not enough free space.