Fix rollback to snapshot when some volumes on which the snapshot was based no longer exist
Previously, we set the ROLLBACK_TARGET flag on the snapshot. Working
with the snapshot like this causes issues when the snapshot exists on
nodes where the resource no longer exists, because the satellite does
not know how to handle the snapshot update.

Now we use a simpler approach where we just set the rollback target as a
resource property, so we don't have to send any snapshot updates to
perform a rollback.