build,pkg: /etc/drbd.d/linstor-resources.res
This adds the res file to the actual packages (deb,rpm). This has the
advantage that we can include a kernel-module dependency there, and
increase it per LINSTOR release while informing the admin in the
appropriate way. The admin can then decide to ignore the dependency or
install the file from the pkg.

LINSTOR should only create this file if it does not exist.