Fix failure when the controller is under load from many peers
Without this fix, the capacity limited queue filled up when more peers
than the number of threads were sending messages and the queue filled up
with messages. Reactor then offered the next value to the queue and
terminated processing because it was not accepted.

When fewer peers were sending many messages this did not cause an issue
because the messages were placed in the unbounded re-ordering queues for
the individual peers. This behavior remains the same.