packaging: bump drbd-utils to >= 9.7.0
With that version we introduced "require-drbd-module-version-*". So far
we don't use such a strict requirement, but be prepared if we eventually
need/want to set it. We might do that pretty soon to get better parser
errors on DRBD8.4 kos. Still, no need to rush that, lets get utils 9.7.0
really out of the door first.