Add option to migrate disk from another node when adding disk to diskless resource
The resource is only removed from the migration source once the sync is
complete, which can take a long time. Hence the API completes once the
disk toggle is complete, and the process is continued by the
'BackgroundRunner'.

The watch for the resource sync remains active while the satellite is
online. Hence it is restarted when the resource is connected rather than
when all resources of the resource definition are connected. This
requires adding a new connection listener method.