Fixed: Exending vlmSize if converted rsc has DRBD type
There was / is a compatibility bug. The old DeviceManager consideres the
VolumeDefinition.size as usable net-size, and enlarges the underlaying
volume with the additionally needed space for metadata.
The new DeviceManager cannot consider the size as net-size, as it is
possible that a new DRBD resource gets deployed on top of an existing
storage resource. In that case the existing size has to be the gross-size
as not all Layers can resize (currently e.g. swordfish).

In future the conroller itself has to take care of writing the gross size
in VolumeDefinition.size in case of deploying multiple layered resource
at once (with policies).