Restore deployment status and error responses
Show the deployment status messages generated by the controller for the
API instead of the generic responses from the satellite. We use a more
robust approach, where merging of the responses from the satellites is
handled outside of the update caller. We no longer rely on getting a
MODIFIED response from the satellite containing the resource name.

Also include the error report ID in the error response.

A higher Reactor version is required for the convenience method
Tuple2.mapT2.