Added LockGuardFactory
LockGuardFactory is a utility class for building LockGuards. The intention
is to solve two major problems we had until now:

1) The order of locks taken and releasing has to match the specification
from the com/linbit/linstor/@LOCK_ORDER. Until now, we had to be careful
at each LockGuard.create* call to have the ordering correct.
LockGuardFactory accepts enums and orders them before creating the actual
LockGuard object. Addtionally, having enums has the convinient side-effect
of not having to inject every single ReadWriteLock to every class, only
the LockGuardFactory.

2) We had many places where we simply forgot to take the reconfiguration-
readLock. This would have became an issue when we would have implemented
a new API and relaying on "if we take the reconfiguration.writeLock, only
we can operate on the whole system".
This issue became obsolte with the LockGuardFactory, as it automatically
adds a reconfiguration-readLock if not specified otherwise. For (not yet
existing) cases where we explicitly do not want any reconfiguraion lock,
the LockGuardFactory can easily be extended with a NO_LOCK enum, instead
of READ or WRITE.