Replaced string based storPool.driver with enum
Also removed most accesses to StorageDriverKind and replaced them with
DeviceProviderKind. The only reason to not delete the deprecated interface
and their implementations is that there are still some "staticTraits"
saved. Storing such information in enums is quite cumbersome. Also we have
to check if thoese information is actually used, otherwise we can simply
remove it.