Fixed hung migrating drbd resource
Resource-ready event was not sent in this case. However, this fix will
send that event twice now (could be optimized), once from the Drbd-Layer
itself and once triggered by the events2 stream (in case of a normal
resource creation. When migrating that second event will be missing, which
was the actual bug)