Replaced DROP and CREATE table with ALTER statements
Making a copy of the old table, dropping the original, creating the
table with original name, but new structure (new or removed columns)
was used as some databases do not support structural transactions.

However, some databases (like postgres) do not allow dropping tables
where foreign keys are pointing to it. This commit therefore replaces
the DROP and reCREATE technique with good old ALTER TABLE statements.
If something fails, the admin should have anyways a copy of the database
before migrating :)