docker: base images on UBI
While this is ways clumsier, we need to base our image on a RHEL product
in order to get certified by OpenShift. This prepares that switch.

For the Makefile we have to move the "uncommitted changes" check to the
non-FORCE version. The RHEL build system likes to add some meta data to
the Dockerfile, which breaks the "uncommitted changes" check. Finally we
switched to propper tarballs anyways, still, it is good to have it
there.

Note that the build-part is based on Centos images. The UBI ones are
very limited in what software you can install. So build in centos, use
the pkgs in the UBI.